Reorganize formatted crate and binaries.
I have now split the binaries in two in the monkey-rs crate.
We have `monkey` for the standard monkey REPL and file reader and the
`monkeyfmt` that handles formatting.
