--- crates/formatter/src/formatter/mod.rs
@@ -1,3 +1,0 @@
-pub mod formatter;
-mod formatter_tests;
-pub mod cli;
