+++ crates/formatter/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod formatter;
-
+pub mod cli;
+mod formatter_tests;
